+++ src/apis/auth/login.api.ts
@@ -1,6 +1,6 @@
-import { LoginRequestDto } from "@/dtos/request/auth/login.request.dto";
+import { LoginRequestDto } from "@/dtos/auth/request/login.request.dto";
 import ResponseDto from "@/dtos/response.dto";
-import LoginResponseDto from "@/dtos/response/auth/login.response.dto";
+import LoginResponseDto from "@/dtos/auth/response/login.response.dto";
 import { axiosInstance, responseErrorHandler, responseSuccessHandler } from "../axiosConfig";
 import { LOGIN_URL } from "../constants";
 import { AxiosError } from "axios";
