+++ src/dtos/auth/response/sign-up-trainer.response.dto.ts
